fix: Move TestingDatabaseInitialzer to framework/testing, add SkipDiscovery and a method to swap database initializer
